Fix genspec patch and output package
The patch to fix the genspec file to accept other org names besides `AWS` and `Alexa` was hard coded for `Generic` but now uses the org name from the target schema. Closes #3

The default package output from `cfn2ts` was specifically build for packages inside the `aws-cdk` project. To support creation of an outside package the python code now creates its own `package.json` and `tsconfig.json` files. Not sure this is 100% correct as my typescript and javascript experience is limited but it seems to work. Closes #4